Move pagination util to fe-kit
